Split up layers.dart file
