+++ lib/routing/views/map.dart
@@ -6,7 +6,9 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:latlong2/latlong.dart';
 import 'package:mapbox_maps_flutter/mapbox_maps_flutter.dart';
-import 'package:priobike/common/map/layers.dart';
+import 'package:priobike/common/map/layers/poi_layers.dart';
+import 'package:priobike/common/map/layers/route_layers.dart';
+import 'package:priobike/common/map/layers/sg_layers.dart';
 import 'package:priobike/common/map/symbols.dart';
 import 'package:priobike/common/map/view.dart';
 import 'package:priobike/positioning/services/positioning.dart';
